requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
websocket-client==1.8.0
zerorpc==0.6.3
